Chore: statusButtonConfig constants로 분리
