+++ co-kkiri/src/components/domains/detail/StatusBasedButton.tsx
@@ -1,37 +1,17 @@
-import Button from "@/components/commons/Button";
+import { useState, useEffect } from "react";
 import styled from "styled-components";
-import { PostApplyStatus } from "@/lib/api/post/type";
-import { ButtonVariant } from "@/components/commons/Button";
 import useOpenToggle from "@/hooks/useOpenToggle";
+import Button from "@/components/commons/Button";
 import ConfirmModal from "@/components/modals/ConfirmModal";
 import InviteResponseModal from "@/components/modals/InviteResponseModal";
+import { statusButtonConfig, StatusButtonConfig } from "@/constants/statusButtonConfig";
 import { ConfirmType } from "@/components/modals/ConfirmModal";
-import { useState, useEffect } from "react";
+import { PostApplyStatus } from "@/lib/api/post/type";
+
 interface MappedButtonProps {
   postApplyStatus: PostApplyStatus;
   className?: string;
 }
-
-type ButtonMapped = {
-  variant: ButtonVariant;
-  text: string;
-  disabled: boolean;
-  type?: ConfirmType;
-};
-
-interface StatusButtonConfig {
-  NOT_APPLIED: ButtonMapped;
-  APPLIED: ButtonMapped;
-  INVITED: ButtonMapped;
-  RECRUIT_CLOSED: ButtonMapped;
-}
-
-const statusButtonConfig: StatusButtonConfig = {
-  NOT_APPLIED: { variant: "primary", text: "지원하기", disabled: false, type: "apply" },
-  APPLIED: { variant: "red", text: "지원 취소", disabled: false, type: "cancel" },
-  INVITED: { variant: "ghost", text: "스카우트 답변하기", disabled: false },
-  RECRUIT_CLOSED: { variant: "primary", text: "모집 완료", disabled: true },
-};
 
 export default function StatusBasedButton({ postApplyStatus, className }: MappedButtonProps) {
   const { isOpen: isConfirmOpen, openToggle: confirmToggle } = useOpenToggle();
